[core] Move ModuleLogger to util package
 * It has nothing to do with native now.
